Layout becomes a function of key labels
Per maintainer review: layout no longer lives as a stored field on
CosmosKeyboard. It's a pure function of the alpha-key labels — like
clusterAngle/clusterSeparation already are. The dropdown calls
detectLayout(\$protoConfig) at render time and applyLayoutToKeys when
the user picks a new option. Custom is the absence of a named match.

This subsumes the maintainer's other code-review concern (the
suspicious reactive `$:` block in VisualEditor2 that broke the
browser back button) — there's nothing to auto-flip because nothing
holds stale state. The dropdown re-derives on every reactivity
trigger.

Schema changes:
- src/proto/cosmos.proto: \`reserved 33;\` for the old layout field
  number. Old URLs with \`layout=N\` are silently dropped on decode;
  the layout is re-derived from the keys' letters.
- CosmosKeyboard interface drops the \`layout\` field; KEYBOARD_DEFAULTS
  drops \`layout\`; encode/decode drops \`encodeLayout\`/\`decodeLayout\`.

Helper relocation: detectLayout, missingKeysFor, alphaColumns,
STANDARD_ALPHA_COLS move from visualEditorHelpers.ts (editor module)
into config.cosmos.ts (the data-model module) so mirrorCluster can
call detectLayout internally without circular imports.

mirrorCluster signature: takes \`kbd: CosmosKeyboard | undefined\`
instead of \`layout: LayoutId\`. Internal callers pass the kbd in
scope; the module-level cluster cache passes \`undefined\` and gets
QWERTY semantics (labels are erased anyway). All 6 internal sites +
3 external sites (HandFitView, toCode, visualEditorHelpers) updated.

Read sites converted to detectLayout(kbd):
- setClusterSize: detect BEFORE the resize so the seed letters reflect
  the user's pre-resize layout. CUSTOM falls back to QWERTY for the
  seed (CUSTOM has no canonical letters).
- Viewer3D Letter input flip-on-read/write: detectLayout(\$protoConfig)
  is the layout for the flipMap.

Write sites removed: the dropdown handler, App.setMode, and the
endToEnd test fixture all stop writing \`kbd.layout\`. The first two
just call applyLayoutToKeys; the test asserts via detectLayout instead.

Tests rewritten to round-trip via the keys (the layout survives
because applyLayoutToKeys writes Colemak/etc. letters into
\`profile.letter\`; serialize/deserialize preserves those; detectLayout
re-derives the named layout from them).
