+++ src/lib/worker/config.cosmos.ts
@@ -2,12 +2,12 @@ import ETrsf, { Constant, fullMirrorETrsf, type MatrixOptions, mirror } from '$l
 // import { deserialize } from 'src/routes/beta/lib/serialize'
 import { flippedKey } from '$lib/geometry/keycaps'
 import { decodeVariant, encodeVariant, PART_INFO, socketSize } from '$lib/geometry/socketsParts'
-import { DEFAULT_LAYOUT, type LayoutId } from '$lib/layouts'
+import { DEFAULT_LAYOUT, flipLetter, getLayout, isAlphaLetter, LAYOUT, type LayoutId, NAMED_LAYOUT_IDS, type NamedLayoutId } from '$lib/layouts'
 import { type ClusterName, type ClusterSide, type ClusterType, type Connector, decodeClusterFlags, encodeClusterFlags, type PartVariant, type ScrewFlags } from '$target/cosmosStructs'
 import type { Curvature } from '$target/proto/cosmos'
 import { Matrix4, Vector3 } from 'three'
 import { type AnyShell, curvature, type Cuttleform, type CuttleKey, decodeTuple, encodeTuple, type FullCuttleform, type Keycap, matrixToRPY, tupleToRot, tupleToXYZ } from './config'
-import { decodePartType, encodePartType, KEYBOARD_DEFAULTS } from './config.serialize'
+import { decodePartType, encodePartType, KEYBOARD_DEFAULTS, LETTERS } from './config.serialize'
 import Trsf from './modeling/transformation'
 import { capitalize, DefaultMap, objEntries, objKeys, sum, TallyMap, trimUndefined } from './util'
 
@@ -111,7 +111,6 @@ export type CosmosKeyboard =
     connectors: ConnectorMaybeCustom[]
     mirrorConnectors: boolean
     plate: Cuttleform['plate']
-    layout: LayoutId
   }
   & ScrewFlags
 
@@ -299,12 +298,12 @@ export function toFullCosmosConfig(conf: FullCuttleform, flipLeft = false): Cosm
   // This halves the URL size and gives better editing experience
   const rightFingers = kbd.clusters.find(c => c.side == 'right' && c.name == 'fingers')
   const leftFingers = kbd.clusters.find(c => c.side == 'left' && c.name == 'fingers')
-  if (rightFingers && leftFingers && stringifyCluster(mirrorCluster(rightFingers, true, kbd.layout)) == stringifyCluster(leftFingers)) {
+  if (rightFingers && leftFingers && stringifyCluster(mirrorCluster(rightFingers, kbd)) == stringifyCluster(leftFingers)) {
     kbd.clusters.splice(kbd.clusters.indexOf(leftFingers), 1)
   }
   const rightThumbs = kbd.clusters.find(c => c.side == 'right' && c.name == 'thumbs')
   const leftThumbs = kbd.clusters.find(c => c.side == 'left' && c.name == 'thumbs')
-  if (rightThumbs && leftThumbs && stringifyCluster(mirrorCluster(rightThumbs, true, kbd.layout)) == stringifyCluster(leftThumbs)) {
+  if (rightThumbs && leftThumbs && stringifyCluster(mirrorCluster(rightThumbs, kbd)) == stringifyCluster(leftThumbs)) {
     kbd.clusters.splice(kbd.clusters.indexOf(leftThumbs), 1)
   }
   sortClusters(kbd.clusters)
@@ -370,7 +369,6 @@ export function toCosmosConfig(conf: Cuttleform, side: 'left' | 'right' | 'unibo
         extensionRight: 8,
       },
     wristRestPosition: overrideWristRest ? KEYBOARD_DEFAULTS.wristRestPosition! : encodeTuple(wrOrigin.xyz().map(t => Math.round(t * 10))),
-    layout: DEFAULT_LAYOUT,
     clusters: side == 'unibody'
       ? [
         ...toCosmosClusters(filterUnibodySide(conf.keys, 'right'), 'right', globalProfile, globalPartType, globalCurvature, wrOriginInv),
@@ -378,7 +376,7 @@ export function toCosmosConfig(conf: Cuttleform, side: 'left' | 'right' | 'unibo
       ]
       : toCosmosClusters(conf.keys, side, globalProfile, globalPartType, globalCurvature, wrOriginInv),
   }
-  if (flipLeft && side == 'left') keyboard.clusters = keyboard.clusters.map(c => mirrorCluster(c, false, keyboard.layout))
+  if (flipLeft && side == 'left') keyboard.clusters = keyboard.clusters.map(c => mirrorCluster(c, keyboard, false))
   sortClusters(keyboard.clusters)
   return keyboard
 }
@@ -483,15 +481,15 @@ export function sideFromCosmosConfig(c: CosmosKeyboard, side: 'left' | 'right' |
     plate: c.plate,
   }
   const clusters: CosmosCluster[] = c.clusters.filter(c => side == 'unibody' || c.side == side)
-  if (side == 'left' && !c.clusters.find(c => c.side == 'left' && c.name == 'fingers')) clusters.unshift(mirrorCluster(c.clusters.find(c => c.side == 'right' && c.name == 'fingers')!, true, c.layout))
+  if (side == 'left' && !c.clusters.find(c => c.side == 'left' && c.name == 'fingers')) clusters.unshift(mirrorCluster(c.clusters.find(c => c.side == 'right' && c.name == 'fingers')!, c))
   if (side == 'unibody' && !c.clusters.find(c => c.side == 'left' && c.name == 'fingers')) {
     clusters.splice(
       2,
       0,
-      mirrorCluster(c.clusters.find(c => c.side == 'right' && c.name == 'fingers')!, true, c.layout),
+      mirrorCluster(c.clusters.find(c => c.side == 'right' && c.name == 'fingers')!, c),
     )
   }
-  if (side != 'right' && !c.clusters.find(c => c.side == 'left' && c.name == 'thumbs')) clusters.push(mirrorCluster(c.clusters.find(c => c.side == 'right' && c.name == 'thumbs')!, true, c.layout))
+  if (side != 'right' && !c.clusters.find(c => c.side == 'left' && c.name == 'thumbs')) clusters.push(mirrorCluster(c.clusters.find(c => c.side == 'right' && c.name == 'thumbs')!, c))
   // console.log('CLUSTERS', clusters)
   for (const clusterA of clusters) {
     for (const clusterB of clusterA.clusters) {
@@ -551,7 +549,7 @@ export function fromCosmosConfig(c: CosmosKeyboard, flipLeft = true): FullCuttle
 export function calcClusterTrsf(cluster: CosmosCluster, keeb: CosmosKeyboard) {
   if (cluster.name == 'fingers') return rotationPositionETrsf(cluster, false)
 
-  const fingerCluster = keeb.clusters.find(c => c.side == cluster.side && c.name == 'fingers') || mirrorCluster(keeb.clusters.find(c => c.name == 'fingers')!, true, keeb.layout)
+  const fingerCluster = keeb.clusters.find(c => c.side == cluster.side && c.name == 'fingers') || mirrorCluster(keeb.clusters.find(c => c.name == 'fingers')!, keeb)
   const fingerPos = rotationPositionETrsf(fingerCluster, false) || new ETrsf()
   const fingerPosInv = rotationPositionInvETrsf(fingerCluster) || new ETrsf()
   return (rotationPositionETrsf(cluster) || new ETrsf()).transformBy(fingerPosInv).transformBy(fingerPos)
@@ -915,7 +913,15 @@ function mirrorRotationTuple(t: bigint | undefined) {
   return encodeTuple([x, -y, -z])
 }
 
-export function mirrorCluster(clr: CosmosCluster, flipLetters = true, layout: LayoutId = DEFAULT_LAYOUT): CosmosCluster {
+/** Synthesize the left side of a (right-side) cluster.
+ *
+ *  `kbd` is needed to derive the layout (via `detectLayout`) — alpha-letter
+ *  flipping uses the layout's flipMap. Pass `undefined` at module-level call
+ *  sites (e.g. cached cluster libraries) where no kbd is in scope; the result
+ *  there falls back to QWERTY semantics, which is fine because those callers
+ *  erase letters before display anyway. */
+export function mirrorCluster(clr: CosmosCluster, kbd: CosmosKeyboard | undefined, flipLetters = true): CosmosCluster {
+  const layout = kbd ? detectLayout(kbd) : DEFAULT_LAYOUT
   return {
     ...clr,
     curvature: { ...clr.curvature },
@@ -962,4 +968,105 @@ export function calculateSplay(column: CosmosCluster, cluster: CosmosCluster) {
 export function nthSplay(conf: CosmosKeyboard, n: number) {
   const { column, cluster } = nthKey(conf, n)
   return calculateSplay(column, cluster)
+}
+
+// --- Layout detection -------------------------------------------------------
+//
+// Layout used to be a stored field on CosmosKeyboard. It's now a pure function
+// of the alpha-key labels — like clusterAngle/clusterSeparation. detectLayout
+// inspects the right finger cluster and reports the named layout it matches,
+// or LAYOUT.CUSTOM. mirrorCluster() calls it internally so callers don't have
+// to thread the layout through.
+
+/** Canonical 5-col alpha block (Q-W-E-R-T row 2, A-S-D-F-G row 3, Z-X-C-V-B
+ *  row 4). All registered layouts share this width; an optional 6th outer-
+ *  pinky column (`'` in QWERTY/Colemak/Workman, `-` in Dvorak) is treated as
+ *  extra and doesn't participate in matching. */
+const STANDARD_ALPHA_COLS = 5
+
+/** Find the indices of the alpha/letter columns within a finger cluster.
+ *  Picks the up-to-5 columns with the most letter-keys; sorts by physical
+ *  column position (left → right). */
+export function alphaColumns(kbd: CosmosKeyboard, cluster: CosmosCluster) {
+  const columns = cluster.clusters.map((col, index) => ({
+    index,
+    column: col.column ?? -1000,
+    nLetters: col.keys
+      .filter(k => PART_INFO[k.partType.type || col.partType.type || kbd.partType.type!].keycap && LETTERS.includes(k.profile.letter!))
+      .length,
+  })).filter(c => c.nLetters > 0)
+  columns.sort((a, b) => b.nLetters - a.nLetters)
+  const topColumns = columns.slice(0, STANDARD_ALPHA_COLS)
+  return topColumns.sort((a, b) => a.column - b.column).map(c => c.index)
+}
+
+/** Inspect the right-finger alpha block and return the named layout it
+ *  matches one-for-one, or `LAYOUT.CUSTOM` if no named layout fits. Match
+ *  rules:
+ *    - At least STANDARD_ALPHA_COLS alpha columns required (else CUSTOM).
+ *    - Positions [0, STANDARD_ALPHA_COLS) on rows 2/3/4 must agree with the
+ *      candidate layout's right-side letters (left clusters compare against
+ *      the layout's flipMap-mirrored letters).
+ *    - Extra alpha columns past the canonical block are ignored. */
+export function detectLayout(kbd: CosmosKeyboard): LayoutId {
+  const cluster = kbd.clusters.find(c => c.name === 'fingers' && c.side === 'right')
+    ?? kbd.clusters.find(c => c.name === 'fingers')
+  if (!cluster) return DEFAULT_LAYOUT
+  const alphas = alphaColumns(kbd, cluster)
+  if (alphas.length < STANDARD_ALPHA_COLS) return LAYOUT.CUSTOM
+  const isLeft = cluster.side === 'left'
+
+  const observed: Record<2 | 3 | 4, (string | undefined)[]> = { 2: [], 3: [], 4: [] }
+  for (let i = 0; i < alphas.length; i++) {
+    const col = cluster.clusters[alphas[i]]
+    const letterCol = isLeft ? alphas.length - 1 - i : i
+    for (const row of [2, 3, 4] as const) {
+      const key = col.keys.find(k => k.profile.row === row && isAlphaLetter(k.profile.letter))
+      observed[row][letterCol] = key?.profile.letter
+    }
+  }
+
+  for (const id of NAMED_LAYOUT_IDS) {
+    const layout = getLayout(id)
+    if (
+      rowMatches(observed[2], layout.rightRows[2], isLeft, id)
+      && rowMatches(observed[3], layout.rightRows[3], isLeft, id)
+      && rowMatches(observed[4], layout.rightRows[4], isLeft, id)
+    ) {
+      return id
+    }
+  }
+  return LAYOUT.CUSTOM
+}
+
+function rowMatches(observed: (string | undefined)[], layoutRow: string, isLeft: boolean, layoutId: NamedLayoutId): boolean {
+  const upper = Math.min(STANDARD_ALPHA_COLS, layoutRow.length)
+  for (let col = 0; col < upper; col++) {
+    const expectedRight = layoutRow.charAt(col) || undefined
+    if (!expectedRight) continue
+    const expected = isLeft ? flipLetter(expectedRight, layoutId) : expectedRight
+    const observedLetter = observed[col]
+    if (!observedLetter || observedLetter !== expected) return false
+  }
+  return true
+}
+
+/** Letters the named layout requires that the keyboard can't currently host
+ *  (because the user removed columns down past STANDARD_ALPHA_COLS). Empty
+ *  when the kbd already has enough alpha columns. */
+export function missingKeysFor(kbd: CosmosKeyboard, target: NamedLayoutId): string[] {
+  const cluster = kbd.clusters.find(c => c.name === 'fingers' && c.side === 'right')
+    ?? kbd.clusters.find(c => c.name === 'fingers')
+  if (!cluster) return []
+  const have = alphaColumns(kbd, cluster).length
+  if (have >= STANDARD_ALPHA_COLS) return []
+  const layout = getLayout(target)
+  const missing: string[] = []
+  for (let col = have; col < STANDARD_ALPHA_COLS; col++) {
+    for (const row of [2, 3, 4] as const) {
+      const letter = layout.rightRows[row].charAt(col)
+      if (letter) missing.push(letter)
+    }
+  }
+  return missing
 }
