+++ src/routes/beta/lib/dialogs/HandFitView.svelte
@@ -93,7 +93,7 @@
             setStaggersCluster(leftSide, staggerLeft, true)
           } else {
             // Mirror the right cluster to the left, then set its size
-            const newLeftSide = mirrorCluster(rightSide!, true, c.layout)
+            const newLeftSide = mirrorCluster(rightSide!, c)
             setStaggersCluster(newLeftSide, staggerLeft, true)
             c.clusters.splice(2, 0, newLeftSide)
           }
