+++ src/routes/beta/lib/editor/visualEditorHelpers.ts
@@ -1,11 +1,13 @@
 import { PART_INFO } from '$lib/geometry/socketsParts'
-import { DEFAULT_LAYOUT, flipLetter, getLayout, isAlphaLetter, LAYOUT, type LayoutId, NAMED_LAYOUT_IDS, type NamedLayoutId, rightSideLetter } from '$lib/layouts'
+import { DEFAULT_LAYOUT, flipLetter, isAlphaLetter, LAYOUT, type LayoutId, rightSideLetter } from '$lib/layouts'
 import { approximateCosmosThumbOrigin, cosmosFingers, type Cuttleform, decodeTuple, encodeTuple, type FullCuttleform, newGeometry } from '$lib/worker/config'
 import {
+  alphaColumns,
   type ConnectorMaybeCustom,
   type CosmosCluster,
   type CosmosKey,
   type CosmosKeyboard,
+  detectLayout,
   fromCosmosConfig,
   indexOfKey,
   mirrorCluster,
@@ -13,7 +15,7 @@ import {
   sideFromCosmosConfig,
   toCosmosConfig,
 } from '$lib/worker/config.cosmos'
-import { decodeCosmosCluster, LETTERS } from '$lib/worker/config.serialize'
+import { decodeCosmosCluster } from '$lib/worker/config.serialize'
 import { estimatedBB } from '$lib/worker/geometry'
 import { Vector } from '$lib/worker/modeling/transformation'
 import { mapObj, objKeys, sum } from '$lib/worker/util'
@@ -43,7 +45,13 @@ export function setClusterSize(keyboard: CosmosKeyboard, side: 'left' | 'right',
   if (side == 'left') newThumb.x *= -1
   const newPosition = originalPosition.add(originalThumb).sub(newThumb)
   const newTup = encodeTuple(newPosition.toArray().map((x) => Math.round(10 * x)))
-  fingers.clusters = cosmosFingers(rows, cols, side, addExtraRow, keyboard.layout)
+  // Detect the kbd's current layout BEFORE we replace the cluster — the new
+  // alpha keys get seeded with that layout's letters. CUSTOM has no rightRows,
+  // so fall back to QWERTY for the seeding (the user's existing letters
+  // outside the alpha block are preserved by the unchanged outer columns).
+  const detected = detectLayout(keyboard)
+  const seedLayout = detected === LAYOUT.CUSTOM ? DEFAULT_LAYOUT : detected
+  fingers.clusters = cosmosFingers(rows, cols, side, addExtraRow, seedLayout)
   fingers.position = newTup
 
   return keyboard
@@ -197,7 +205,7 @@ export function setThumbCluster(c: CosmosKeyboard, type: Thumb, side: 'left' | '
   }
   // Erase all key labels
   cluster.clusters.forEach(c => c.keys.forEach(k => k.profile.letter = undefined))
-  if (side == 'left') cluster = mirrorCluster(cluster, true, c.layout)
+  if (side == 'left') cluster = mirrorCluster(cluster, c)
 
   const thumb = c.clusters.find((c) => c.name == 'thumbs' && c.side == side)
   if (!thumb) return c
@@ -210,7 +218,7 @@ export function setThumbCluster(c: CosmosKeyboard, type: Thumb, side: 'left' | '
 }
 
 const decodedClusters = mapObj(THUMB_CONFIG, ({ b64 }) => decodeCosmosCluster(Cluster.fromBinary(Uint8Array.from(atob(b64), c => c.charCodeAt(0)))))
-const mirroredDecodedClusters = mapObj(decodedClusters, c => mirrorCluster(c))
+const mirroredDecodedClusters = mapObj(decodedClusters, c => mirrorCluster(c, undefined))
 
 export function getThumbN(c: CosmosKeyboard, side: 'left' | 'right') {
   const thumb = c.clusters.find((c) => c.name == 'thumbs' && c.side == side)
@@ -321,23 +329,6 @@ export function filterKeys(kbd: CosmosKeyboard, predicate: (k: CosmosKey, col: C
 export function mapKeys<T>(kbd: CosmosKeyboard, predicate: (k: CosmosKey, col: CosmosCluster, cl: CosmosCluster, index: number) => T): T[] {
   let i = 0
   return kbd.clusters.flatMap(cluster => cluster.clusters.flatMap(col => col.keys.map(k => predicate(k, col, cluster, i++))))
-}
-
-/** Find the indices of the five alpha/letter columns.
- * These are determined by finding the 5 columns with the largest number of letters.
- * They are sorted by column position from smallest to largest
- */
-function alphaColumns(kbd: CosmosKeyboard, cluster: CosmosCluster) {
-  const columns = cluster.clusters.map((col, index) => ({
-    index,
-    column: col.column ?? -1000,
-    nLetters: col.keys
-      .filter(k => PART_INFO[k.partType.type || col.partType.type || kbd.partType.type!].keycap && LETTERS.includes(k.profile.letter!))
-      .length,
-  })).filter(c => c.nLetters > 0)
-  columns.sort((a, b) => b.nLetters - a.nLetters)
-  const topColumns = columns.slice(0, 5)
-  return topColumns.sort((a, b) => a.column - b.column).map(c => c.index)
 }
 
 export function addRow(kbd: CosmosKeyboard, fn: (side: 'left' | 'right', alphas: number[], row: number, column: number) => string | null) {
@@ -515,101 +506,4 @@ export function applyLayoutToKeys(kbd: CosmosKeyboard, layoutId: LayoutId): Cosm
       }
     })
   })
-}
-
-/**
- * Canonical width of the alpha block (Q-W-E-R-T on row 2, A-S-D-F-G on row 3,
- * Z-X-C-V-B on row 4 for QWERTY). All registered layouts share this 5-col
- * core; the 6th column on row 3 (`'` in QWERTY/Colemak/Workman, `-` in Dvorak)
- * is an optional outer-pinky punctuation that some keyboards include.
- *
- * detectLayout matches against the first STANDARD_ALPHA_COLS positions and
- * ignores extras in either direction. missingKeysFor uses the same width to
- * compute what the user needs to add to host a named layout.
- */
-const STANDARD_ALPHA_COLS = 5
-
-/**
- * Inspect `kbd`'s right-finger alpha block and return the named layout it
- * matches one-for-one, or `LAYOUT.CUSTOM` if no named layout fits.
- *
- * Match rules:
- *   - The keyboard must have at least STANDARD_ALPHA_COLS alpha columns —
- *     fewer can't host any layout's full alpha block, so the answer is CUSTOM.
- *   - Positions [0, STANDARD_ALPHA_COLS) on rows 2/3/4 must agree with the
- *     candidate layout's right-side letters (left-side keyboards compare
- *     against the layout's flipMap-mirrored letters).
- *   - Extra alpha columns beyond the canonical block are ignored — adding
- *     keys "outside the layout's range" shouldn't flip the dropdown.
- *
- * Used by:
- *   - the expert→basic mode transition (so the layout dropdown reflects the
- *     keymap the user actually has, instead of always reverting to QWERTY)
- *   - the "did the user start customizing?" auto-flip to CUSTOM after a
- *     legend edit or column delete
- */
-export function detectLayout(kbd: CosmosKeyboard): LayoutId {
-  const cluster = kbd.clusters.find(c => c.name === 'fingers' && c.side === 'right')
-    ?? kbd.clusters.find(c => c.name === 'fingers')
-  if (!cluster) return DEFAULT_LAYOUT
-  const alphas = alphaColumns(kbd, cluster)
-  if (alphas.length < STANDARD_ALPHA_COLS) return LAYOUT.CUSTOM
-  const isLeft = cluster.side === 'left'
-
-  const observed: Record<2 | 3 | 4, (string | undefined)[]> = { 2: [], 3: [], 4: [] }
-  for (let i = 0; i < alphas.length; i++) {
-    const col = cluster.clusters[alphas[i]]
-    const letterCol = isLeft ? alphas.length - 1 - i : i
-    for (const row of [2, 3, 4] as const) {
-      const key = col.keys.find(k => k.profile.row === row && isAlphaLetter(k.profile.letter))
-      observed[row][letterCol] = key?.profile.letter
-    }
-  }
-
-  for (const id of NAMED_LAYOUT_IDS) {
-    const layout = getLayout(id)
-    if (
-      rowMatches(observed[2], layout.rightRows[2], isLeft, id)
-      && rowMatches(observed[3], layout.rightRows[3], isLeft, id)
-      && rowMatches(observed[4], layout.rightRows[4], isLeft, id)
-    ) {
-      return id
-    }
-  }
-  return LAYOUT.CUSTOM
-}
-
-function rowMatches(observed: (string | undefined)[], layoutRow: string, isLeft: boolean, layoutId: NamedLayoutId): boolean {
-  const upper = Math.min(STANDARD_ALPHA_COLS, layoutRow.length)
-  for (let col = 0; col < upper; col++) {
-    const expectedRight = layoutRow.charAt(col) || undefined
-    if (!expectedRight) continue
-    const expected = isLeft ? flipLetter(expectedRight, layoutId) : expectedRight
-    const observedLetter = observed[col]
-    if (!observedLetter || observedLetter !== expected) return false
-  }
-  return true
-}
-
-/**
- * Letters the named layout requires that the keyboard's right-finger alpha
- * block can't currently host (because the user removed columns down past the
- * canonical 5-col block). Empty when the keyboard has at least
- * STANDARD_ALPHA_COLS alpha columns.
- */
-export function missingKeysFor(kbd: CosmosKeyboard, target: NamedLayoutId): string[] {
-  const cluster = kbd.clusters.find(c => c.name === 'fingers' && c.side === 'right')
-    ?? kbd.clusters.find(c => c.name === 'fingers')
-  if (!cluster) return []
-  const have = alphaColumns(kbd, cluster).length
-  if (have >= STANDARD_ALPHA_COLS) return []
-  const layout = getLayout(target)
-  const missing: string[] = []
-  for (let col = have; col < STANDARD_ALPHA_COLS; col++) {
-    for (const row of [2, 3, 4] as const) {
-      const letter = layout.rightRows[row].charAt(col)
-      if (letter) missing.push(letter)
-    }
-  }
-  return missing
 }
